requirements.txt
# --- Core ML / Embedding Stack ---
torch>=2.1.0,<3.0         # Stable with transformers and sentence-transformers
transformers>=4.40.0,<5.0
sentence-transformers>=3.0.1,<6.0
numpy>=1.23.0,<3.0
scikit-learn>=1.3.0,<2.0
scipy>=1.11.0,<2.0
pandas>=2.0.0,<3.0
tqdm>=4.66.0

# --- Haystack Ecosystem ---
haystack-ai==2.19.0
chroma-haystack>=3.0.0
chromadb>=1.2.0,<2.0.0

# --- Code Analysis & Metrics ---
radon>=6.0.1
pylint>=3.0.0,<5.0
GitPython==3.1.45
rbo==0.1.3 
networkx>=3.3,<4.0
matplotlib>=3.8.0,<4.0
seaborn>=0.13.0,<1.0

# --- Static Analysis / AST Tools ---
astroid>=3.1.0,<5.0
PyYAML>=6.0,<7.0
nltk>=3.9.1  # for Haystack DocumentSplitter
PyMuPDF>=1.24.0,<2.0  # correct package for `import fitz`

# --- Data / NLP Utilities ---
datasets>=2.19.0,<3.0
huggingface-hub>=0.24.0,<1.0
codecarbon>=3.0.8,<4.0

# --- Optional / Interface Tools ---
gradio>=5.0.0,<6.0
requests>=2.31.0,<3.0

# --- Testing & Utilities ---
pytest>=8.0.0,<9.0
setuptools>=68.0.0

# --- Compatibility Fixes ---
importlib-metadata>=7.0.0,<9.0

# --- Lang chain ---
langchain-text-splitters==1.0.0
langchain==1.1.2

langchain-ollama==1.0.0